Frontend tests pass
